Decomposed client
